Improve the error message when the build data file has been deleted (#1963)
On Honeycomb I've seen cases of builds failing due to a pre/post-compile
hook having deleted the buildpack's internal build data file - typically
as part of the user trying to clean up the build cache.

This file is required for the buildpack to function correctly, so
deleting it isn't supported. However, we can show a clearer error
message.

Before:

```
-----> Running bin/post_compile hook
       + rm -rf /tmp/cache
jq: error: Could not open file /tmp/cache/build-data/python.json: No such file or directory
jq: error: Could not open file /tmp/cache/build-data/python.json: No such file or directory

 !     Internal Error: An unhandled buildpack error occurred.
 !
 !     An unhandled error occurred while executing the command:
 !     jq --exit-status --arg key "${key}" "${jq_args[@]}" '. + { ($key): ($value) }' "${BUILD_DATA_FILE}"
 !
 !     If this issue persists, please open a support ticket or file
 !     an issue on the buildpack's GitHub repository:
 !     https://help.heroku.com/
 !     https://github.com/heroku/heroku-buildpack-python/issues
 !
 !     Stack trace:
 !     build_data::_set @ /tmp/buildpack/lib/build_data.sh:109
 !     build_data::set_raw @ /tmp/buildpack/lib/build_data.sh:76
 !     build_data::set_duration @ /tmp/buildpack/lib/build_data.sh:60
 !     hooks::run_hook @ /tmp/buildpack/lib/hooks.sh:37
 !     main @ ./bin/compile:289

jq: error: Could not open file /tmp/cache/build-data/python.json: No such file or directory
jq: error: Could not open file /tmp/cache/build-data/python.json: No such file or directory

 !     Internal Error: An unhandled buildpack error occurred.
 !
 !     An unhandled error occurred while executing the command:
 !     new_data_file_contents="$(jq --exit-status --arg key "${key}" "${jq_args[@]}" '. + { ($key): ($value) }' "${BUILD_DATA_FILE}")"
 !
 !     If this issue persists, please open a support ticket or file
 !     an issue on the buildpack's GitHub repository:
 !     https://help.heroku.com/
 !     https://github.com/heroku/heroku-buildpack-python/issues
 !
 !     Stack trace:
 !     build_data::_set @ /tmp/buildpack/lib/build_data.sh:109
 !     build_data::set_raw @ /tmp/buildpack/lib/build_data.sh:76
 !     build_data::set_duration @ /tmp/buildpack/lib/build_data.sh:60
 !     hooks::run_hook @ /tmp/buildpack/lib/hooks.sh:37
 !     main @ ./bin/compile:289

jq: error: Could not open file /tmp/cache/build-data/python.json: No such file or directory
```

After:

```
-----> Running bin/post_compile hook
       + rm -rf /tmp/cache

 !     Error: The buildpack's build data file has been deleted.
 !
 !     The Python buildpack's build data file can't be found:
 !     /tmp/cache/build-data/python.json
 !
 !     This file is required for the buildpack to work correctly,
 !     and so you must not delete it when removing files from the
 !     build cache or /tmp directories.
```

Lastly, we also now use `--exit-status` in more places to prevent
invalid empty build data files from being silently ignored.

GUS-W-20220731.

+++ lib/build_data.sh
@@ -104,9 +104,25 @@ function build_data::_set() {
 		local jq_args=(--argjson value "${value}")
 	fi
 
-	local new_data_file_contents
-	new_data_file_contents="$(jq --arg key "${key}" "${jq_args[@]}" '. + { ($key): ($value) }' "${BUILD_DATA_FILE}")"
-	echo "${new_data_file_contents}" >"${BUILD_DATA_FILE}"
+	if [[ -f "${BUILD_DATA_FILE}" ]]; then
+		local new_data_file_contents
+		new_data_file_contents="$(jq --exit-status --arg key "${key}" "${jq_args[@]}" '. + { ($key): ($value) }' "${BUILD_DATA_FILE}")"
+		echo "${new_data_file_contents}" >"${BUILD_DATA_FILE}"
+	else
+		output::error <<-EOF
+			Error: Can't find the buildpack's build data file.
+
+			The Python buildpack's internal build data file is missing:
+			${BUILD_DATA_FILE}
+
+			This file is required for the buildpack to work correctly,
+			and so you must not delete it when removing files from the
+			build cache or /tmp directories.
+		EOF
+		build_data::setup
+		build_data::set_string "failure_reason" "build-data::data-file-deleted"
+		exit 1
+	fi
 }
 
 # Check whether an entry exists in the build data store for the current build.
@@ -143,6 +159,7 @@ function build_data::get_previous() {
 		tac "${LEGACY_BUILD_DATA_FILE}" | { grep --perl-regexp --only-matching --max-count=1 "^${key}=\K.*$" || true; }
 	elif [[ -f "${PREVIOUS_BUILD_DATA_FILE}" ]]; then
 		# The `// empty` ensures we return the empty string rather than `null` if the key doesn't exist.
+		# We don't use `--exit-status` since `false` is a valid value for us to retrieve.
 		jq --raw-output ".${key} // empty" "${PREVIOUS_BUILD_DATA_FILE}"
 	fi
 }
@@ -170,5 +187,5 @@ function build_data::current_unix_realtime() {
 # build_data::print_bin_report_json
 # ```
 function build_data::print_bin_report_json() {
-	jq --sort-keys '.' "${BUILD_DATA_FILE}"
+	jq --exit-status --sort-keys '.' "${BUILD_DATA_FILE}"
 }
